Many bug fixes and improvements (v2.2.1)
- add support for SiFive E21 (4 PMP regs / SiFive CLIC)
- increase CLIC max sources to 256
- zone1: add printf.c and strtok.c to reduce code size < 16K
- zone1: switch to vectored trap handler to reduce irq latency
- fix -Wall and -Wextra warnings
- update riscv-gnu-toolchain-20210618 (rv32i|rv32e|rv64i)mac
- update manual.pdf and README.md

multizone.jar @1b9f31 (v2.2.1) / kernel @6b58e8 (v2.2.1):

- major redesign of kernel internal data structures
- up to 32/64 clint + 64/128 plic sources
- up to 256 clic sources for both rv32 and rv64
- reduce size of most kern vars from REGSIZE to 4-byte
- increase trap & emulation i-cache size
- add support for SiFive clic (E21)
- reintroduce support for rv32e 16-regs
- update riscv-gnu-toolchain-20210618 (rv32i|rv32e|rv64i)c

+++ zone3/main.c
@@ -82,7 +82,7 @@ void timer_set(const int i, const uint64_t timecmp){
 	timer[i].timecmp = timecmp;
 
 	uint64_t timecmp_min = UINT64_MAX;
-	for (int i=0; i<sizeof(timer)/sizeof(timer[0]); i++)
+	for (size_t i=0; i<sizeof(timer)/sizeof(timer[0]); i++)
 		timecmp_min = timer[i].timecmp < timecmp_min ? timer[i].timecmp : timecmp_min;
 
 	MZONE_WRTIMECMP(timecmp_min);
@@ -90,7 +90,7 @@ void timer_set(const int i, const uint64_t timecmp){
 }
 void timer_handler(const uint64_t time){
 
-	for (int i=0; i<sizeof(timer)/sizeof(timer[0]); i++)
+	for (size_t i=0; i<sizeof(timer)/sizeof(timer[0]); i++)
 		if(time >= timer[i].timecmp){
 			timer_set(i, timer[i].task());
 			break;
